change repository name.

+++ install.sh
@@ -1,7 +1,10 @@
 #!/bin/sh
 
 for f in .[A-z]*; do
+    if [ $f != ".git" ]
+    then
 	rm -f ../$f
-	ln -s ./rcfiles/$f $HOME/$f
+	ln -s ./myrcfiles/$f $HOME/$f
+    fi
 done
